+++ src/editor/index.js
@@ -68,6 +68,21 @@ addFilter( 'blocks.registerBlockType', 'newspack-newsletters/core-blocks', ( set
 	return settings;
 } );
 
+if ( newspack_email_editor_data.supported_social_icon_services ) {
+	addFilter(
+		'blocks.registerBlockType',
+		'newspack-newsletters/core-social-links',
+		( settings, name ) => {
+			if ( 'core/social-link' === name && settings.variations ) {
+				settings.variations = settings.variations.filter( variation =>
+					newspack_email_editor_data.supported_social_icon_services.includes( variation.name )
+				);
+			}
+			return settings;
+		}
+	);
+}
+
 registerBlockStyle( 'core/social-links', {
 	name: 'circle-black',
 	label: __( 'Circle Black', 'newspack-newsletters' ),